requirements.txt
grpcio
grpcio-tools
confluent_kafka
elasticsearch